Add lecture files and import arrayTypes
