+++ src/app.ts
@@ -1,18 +1,3 @@
-console.log("app.js loaded");
-
-type Person = {
-  name: string;
-  age: number;
-};
-
-const person: Person = {
-  name: "Jerome",
-  age: 36,
-};
-
-// const person: { name: string; age: number } = {
-//   name: "Jerome",
-//   age: 36,
-// };
-
-console.log(person.name);
+//import './lectures/basics';
+//import './lectures/objectTypes';
+import './lectures/arrayTypes'
