+++ src/lectures/basics.ts
