+++ src/lectures/objectTypes.ts
@@ -1,0 +1,18 @@
+console.log("app.js loaded");
+
+type Person = {
+  name: string;
+  age: number;
+};
+
+const person: Person = {
+  name: "Jerome",
+  age: 36,
+};
+
+// const person: { name: string; age: number } = {
+//   name: "Jerome",
+//   age: 36,
+// };
+
+console.log(person.name);
